# src/main/resources/api/api.yml
openapi: 3.0.3
info:
  version: '1.0'
  title: Company Management API
  description: |
    API CRUD pour la gestion des entreprises, chantiers, entrepôts, équipements, matériaux,
    mouvements financiers, dépenses, achats et frais de transport.
    Basée sur le schéma de base de données fourni.
servers:
  - url: 'https://api-dev.company.com'
  - url: 'https://api-prod.company.com'
security:
  - BearerAuth: [ ]
paths:
...
  '/users/{id}':
    get:
      tags:
        - Users
      summary: Get user by identifier
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - Users
      summary: Delete a user by identifier
      operationId: deleteUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /users:
    get:
      tags:
        - Users
      summary: Get all users
      operationId: getUsers
      parameters:
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: first_name
          in: query
          description: "Filter users by first name, case is ignored"
          schema:
            type: string
        - name: last_name
          in: query
          description: "Filter users by last name, case is ignored"
          schema:
            type: string
        - name: email
          in: query
          description: "Filter users by email, case is ignored"
          schema:
            type: string
        - name: role
          in: query
          description: "Filter users by role"
          schema:
            $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - Users
      summary: Create new users or update existing users
      description: Update users when `id` are provided, create them otherwise.
      operationId: crupdateUsers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CrupdateUser'
      responses:
        '200':
          description: The created or updated users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'



  # =========================
  # COMPANY
  # =========================
...
  '/monetary_movements/{id}':
    get:
      tags:
        - MonetaryMovement
      summary: Get monetary movement by identifier
      operationId: getMonetaryMovementById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified monetary movement
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonetaryMovement'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - MonetaryMovement
      summary: Delete monetary movement by identifier
      operationId: deleteMonetaryMovementById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Monetary movement deleted successfully
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /monetary_movements:
    get:
      tags:
        - MonetaryMovement
      summary: Get all monetary movements
      operationId: getMonetaryMovements
      parameters:
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
      responses:
        '200':
          description: List of monetary movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MonetaryMovement'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - MonetaryMovement
      summary: Create new monetary movements or update existing ones
      operationId: crupdateMonetaryMovements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CrupdateMonetaryMovement'
      responses:
        '200':
          description: The created or updated monetary movements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MonetaryMovement'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'


  # =========================
  # EXPENSE
  # =========================
...
  '/companies/{comp_id}/job/{job_id}/user/{user_id}/expenses/{expenses_id}/employee_payments/{id}':
    get:
      tags:
        - EmployeePayment
      summary: Get employee payment by identifier
      operationId: getEmployeePaymentById
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: string
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The identified employee payment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeePayment'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    delete:
      tags:
        - EmployeePayment
      summary: Delete employee payment by identifier
      operationId: deleteEmployeePaymentById
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: string
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Employee payment deleted successfully
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
  /companies/{comp_id}/job/{job_id}/user/{user_id}/expenses/{expenses_id}/employee_payments:
    get:
      tags:
        - EmployeePayment
      summary: Get all employee payments
      operationId: getEmployeePayments
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: string
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: expenses_id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            $ref: '#/components/schemas/Page'
        - name: page_size
          in: query
          schema:
            $ref: '#/components/schemas/PageSize'
        - name: employee_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of employee payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeePayment'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'
    put:
      tags:
        - EmployeePayment
      summary: Create new employee payments or update existing ones
      operationId: crupdateEmployeePayments
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: string
        - name: job_id
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: path
          required: true
          schema:
            type: string
        - name: expenses_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/CrupdateEmployeePayment'
      responses:
        '200':
          description: The created or updated employee payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeePayment'
        '400':
          $ref: '#/components/responses/400'
        '403':
          $ref: '#/components/responses/403'
        '404':
          $ref: '#/components/responses/404'
        '429':
          $ref: '#/components/responses/429'
        '500':
          $ref: '#/components/responses/500'


  # =========================
  # TRAVEL PEOPLE
  # =========================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  responses:
    '400':
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestException'
    '403':
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotAuthorizedException'
    '404':
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ResourceNotFoundException'
    '429':
      description: Too many requests to the API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/TooManyRequestsException'
    '500':
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerException'
  schemas:
...
    PaymentType:
      type: string
      enum:
        - ADVANCE
        - MONTHLY
        - OTHER
...
    Sex:
      type: string
      enum:
        - M
        - F
    Role:
      type: string
      enum:
        - ADMIN
        - WAREHOUSE_WORKER
        - EMPLOYEE
        - ADMINISTRATION

    # =========================
    # COMMON
    # =========================
    AuditFields:
      properties:
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
        updated_by:
          type: string
        comment:
          type: string

    # =========================
    # USER
    # =========================
    User:
      allOf:
        - $ref: '#/components/schemas/AuditFields'
      properties:
        id:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        first_name:
          type: string
        last_name:
          type: string
        sex:
          $ref: '#/components/schemas/Sex'
        email:
          type: string
    CrupdateUser:
      properties:
        id:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        first_name:
          type: string
        last_name:
          type: string
        sex:
          $ref: '#/components/schemas/Sex'
        email:
          type: string
        password:
          type: string

    # =========================
    # COMPANY
    # =========================
...
    MonetaryMovement:
      allOf:
        - $ref: '#/components/schemas/AuditFields'
      properties:
        id:
          type: string
        amount:
          type: integer
        description:
          type: string
    CrupdateMonetaryMovement:
      properties:
        id:
          type: string
        amount:
          type: integer
        description:
          type: string

    # =========================
    # INCOME
    # =========================
...
    EmployeePayment:
      allOf:
        - $ref: '#/components/schemas/AuditFields'
      properties:
        id:
          type: string
        expense_id:
          type: string
        employee_id:
          type: string
        payment_description:
          type: string
        payment_type:
          $ref: '#/components/schemas/PaymentType'
    CrupdateEmployeePayment:
      properties:
        id:
          type: string
        expense_id:
          type: string
        employee_id:
          type: string
        payment_description:
          type: string
        payment_type:
          $ref: '#/components/schemas/PaymentType'

    # =========================
    # TRAVEL EXPENSE
    # =========================
...
    Page:
      type: integer
      description: Start at `1`
    PageSize:
      type: integer
      description: Cannot exceed `500`
    Exception:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
    BadRequestException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: BadRequestException
        message: Bad request
    NotAuthorizedException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: NotAuthorizedException
        message: Not authorized
    ResourceNotFoundException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: ResourceNotFoundException
        message: Resource of type <T> identified by <I> not found
    TooManyRequestsException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: TooManyRequestsException
        message: Too many requests
    InternalServerException:
      allOf:
        - $ref: '#/components/schemas/Exception'
      example:
        type: InternalServerException
        message: Unexpected error